requirements.txt
# ERCOT Battery Storage Revenue Opportunity Dashboard
# Zentus - Engie Urja AI Challenge 2025

# Core dashboard framework
streamlit>=1.32.0  # Required for Altair v5 support
streamlit-calendar
protobuf==3.20.3  # Pin for Streamlit compatibility

# Data manipulation
pandas>=2.0.0  # Core data manipulation
polars>=0.19.0  # Fast parquet I/O in ParquetDataLoader (5-10x faster than pandas)
numpy>=1.24.0

# Scientific computing
scipy>=1.10.0  # Linear optimization (linprog) and signal processing

# Visualization
altair>=5.0.0  # Declarative statistical visualizations (AMR optimization charts)
plotly>=5.17.0  # Interactive charts (used in ui/components/charts.py)
matplotlib>=3.5.0  # Required by pandas.style.background_gradient() for colormaps
tqdm>=4.65.0  # Progress bars (fetch_ercot_data.py, fetch_ercot_api.py)

# API and HTTP
requests>=2.28.0  # HTTP requests
gridstatus>=0.24.0  # ERCOT market data access via API (limited to recent data)
urllib3>=1.26.0  # HTTP library (dependency)

# Database
supabase>=2.0.0  # Supabase client for database integration
python-dotenv>=1.0.0  # Environment variable management (loads .env file)